sge_test_backend/requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
sqlalchemy==2.0.41
alembic==1.13.1
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.2.1
beautifulsoup4==4.12.3
requests==2.31.0
apscheduler==3.10.4
pytest==8.0.2
httpx==0.27.0
python-multipart==0.0.9
email-validator==2.1.1
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9 